keyspaces/requirements.txt
cassandra-driver==3.25.0
cassandra-sigv4==4.0.2